requirements.txt
quart==0.18.2
python-telegram-bot==20.6
redis==5.0.1
hypercorn==0.14.2